requirements.txt
cryptography
python-dotenv
flask
flask_cors